+++ go/api/base_client.go
@@ -15,6 +15,7 @@ import (
 	"strconv"
 	"unsafe"
 
+	"github.com/valkey-io/valkey-glide/go/glide/api/options"
 	"github.com/valkey-io/valkey-glide/go/glide/protobuf"
 	"github.com/valkey-io/valkey-glide/go/glide/utils"
 	"google.golang.org/protobuf/proto"
@@ -26,6 +27,7 @@ type BaseClient interface {
 	HashCommands
 	ListCommands
 	SetCommands
+	SortedSetCommands
 	ConnectionManagementCommands
 	GenericBaseCommands
 	// Close terminates the client by closing all associated resources.
@@ -461,6 +463,24 @@ func (client *baseClient) HStrLen(key string, field string) (Result[int64], erro
 	return handleLongResponse(result)
 }
 
+func (client *baseClient) HIncrBy(key string, field string, increment int64) (Result[int64], error) {
+	result, err := client.executeCommand(C.HIncrBy, []string{key, field, utils.IntToString(increment)})
+	if err != nil {
+		return CreateNilInt64Result(), err
+	}
+
+	return handleLongResponse(result)
+}
+
+func (client *baseClient) HIncrByFloat(key string, field string, increment float64) (Result[float64], error) {
+	result, err := client.executeCommand(C.HIncrByFloat, []string{key, field, utils.FloatToString(increment)})
+	if err != nil {
+		return CreateNilFloat64Result(), err
+	}
+
+	return handleDoubleResponse(result)
+}
+
 func (client *baseClient) LPush(key string, elements []string) (Result[int64], error) {
 	result, err := client.executeCommand(C.LPush, append([]string{key}, elements...))
 	if err != nil {
@@ -687,6 +707,39 @@ func (client *baseClient) SUnion(keys []string) (map[Result[string]]struct{}, er
 	return handleStringSetResponse(result)
 }
 
+func (client *baseClient) SScan(key string, cursor string) (Result[string], []Result[string], error) {
+	result, err := client.executeCommand(C.SScan, []string{key, cursor})
+	if err != nil {
+		return CreateNilStringResult(), nil, err
+	}
+	return handleScanResponse(result)
+}
+
+func (client *baseClient) SScanWithOptions(
+	key string,
+	cursor string,
+	options *BaseScanOptions,
+) (Result[string], []Result[string], error) {
+	optionArgs, err := options.toArgs()
+	if err != nil {
+		return CreateNilStringResult(), nil, err
+	}
+
+	result, err := client.executeCommand(C.SScan, append([]string{key, cursor}, optionArgs...))
+	if err != nil {
+		return CreateNilStringResult(), nil, err
+	}
+	return handleScanResponse(result)
+}
+
+func (client *baseClient) SMove(source string, destination string, member string) (Result[bool], error) {
+	result, err := client.executeCommand(C.SMove, []string{source, destination, member})
+	if err != nil {
+		return CreateNilBoolResult(), err
+	}
+	return handleBooleanResponse(result)
+}
+
 func (client *baseClient) LRange(key string, start int64, end int64) ([]Result[string], error) {
 	result, err := client.executeCommand(C.LRange, []string{key, utils.IntToString(start), utils.IntToString(end)})
 	if err != nil {
@@ -761,7 +814,10 @@ func (client *baseClient) LInsert(
 		return CreateNilInt64Result(), err
 	}
 
-	result, err := client.executeCommand(C.LInsert, []string{key, insertPositionStr, pivot, element})
+	result, err := client.executeCommand(
+		C.LInsert,
+		[]string{key, insertPositionStr, pivot, element},
+	)
 	if err != nil {
 		return CreateNilInt64Result(), err
 	}
@@ -1185,4 +1241,90 @@ func (client *baseClient) Renamenx(key string, newKey string) (Result[bool], err
 		return CreateNilBoolResult(), err
 	}
 	return handleBooleanResponse(result)
+}
+
+func (client *baseClient) ZAdd(
+	key string,
+	membersScoreMap map[string]float64,
+) (Result[int64], error) {
+	result, err := client.executeCommand(
+		C.ZAdd,
+		append([]string{key}, utils.ConvertMapToValueKeyStringArray(membersScoreMap)...),
+	)
+	if err != nil {
+		return CreateNilInt64Result(), err
+	}
+
+	return handleLongResponse(result)
+}
+
+func (client *baseClient) ZAddWithOptions(
+	key string,
+	membersScoreMap map[string]float64,
+	opts *options.ZAddOptions,
+) (Result[int64], error) {
+	optionArgs, err := opts.ToArgs()
+	if err != nil {
+		return CreateNilInt64Result(), err
+	}
+	commandArgs := append([]string{key}, optionArgs...)
+	result, err := client.executeCommand(
+		C.ZAdd,
+		append(commandArgs, utils.ConvertMapToValueKeyStringArray(membersScoreMap)...),
+	)
+	if err != nil {
+		return CreateNilInt64Result(), err
+	}
+
+	return handleLongResponse(result)
+}
+
+func (client *baseClient) zAddIncrBase(key string, opts *options.ZAddOptions) (Result[float64], error) {
+	optionArgs, err := opts.ToArgs()
+	if err != nil {
+		return CreateNilFloat64Result(), err
+	}
+
+	result, err := client.executeCommand(C.ZAdd, append([]string{key}, optionArgs...))
+	if err != nil {
+		return CreateNilFloat64Result(), err
+	}
+
+	return handleDoubleResponse(result)
+}
+
+func (client *baseClient) ZAddIncr(
+	key string,
+	member string,
+	increment float64,
+) (Result[float64], error) {
+	options, err := options.NewZAddOptionsBuilder().SetIncr(true, increment, member)
+	if err != nil {
+		return CreateNilFloat64Result(), err
+	}
+
+	return client.zAddIncrBase(key, options)
+}
+
+func (client *baseClient) ZAddIncrWithOptions(
+	key string,
+	member string,
+	increment float64,
+	opts *options.ZAddOptions,
+) (Result[float64], error) {
+	incrOpts, err := opts.SetIncr(true, increment, member)
+	if err != nil {
+		return CreateNilFloat64Result(), err
+	}
+
+	return client.zAddIncrBase(key, incrOpts)
+}
+
+func (client *baseClient) ZIncrBy(key string, increment float64, member string) (Result[float64], error) {
+	result, err := client.executeCommand(C.ZIncrBy, []string{key, utils.FloatToString(increment), member})
+	if err != nil {
+		return CreateNilFloat64Result(), err
+	}
+
+	return handleDoubleResponse(result)
 }