+++ go/api/glide_cluster_client.go
@@ -75,3 +75,41 @@ func (client *GlideClusterClient) PingWithRouteAndMessage(message string, route 
 
 	return handleStringResponse(result)
 }
+
+// CustomCommand executes a single command, specified by args, without checking inputs. Every part of the command, including
+// the command name and subcommands, should be added as a separate value in args. The returning value depends on the executed
+// command.
+//
+// The command will be routed automatically based on the passed command's default request policy.
+//
+// See [Valkey GLIDE Wiki] for details on the restrictions and limitations of the custom command API.
+//
+// This function should only be used for single-response commands. Commands that don't return complete response and awaits
+// (such as SUBSCRIBE), or that return potentially more than a single response (such as XREAD), or that change the client's
+// behavior (such as entering pub/sub mode on RESP2 connections) shouldn't be called using this function.
+//
+// Parameters:
+//
+//	args - Arguments for the custom command including the command name.
+//
+// Return value:
+//
+//	The returned value for the custom command.
+//
+// For example:
+//
+//	result, err := client.CustomCommand([]string{"ping"})
+//	result.Value().(string): "PONG"
+//
+// [Valkey GLIDE Wiki]: https://github.com/valkey-io/valkey-glide/wiki/General-Concepts#custom-command
+func (client *GlideClusterClient) CustomCommand(args []string) (ClusterValue[interface{}], error) {
+	res, err := client.executeCommand(C.CustomCommand, args)
+	if err != nil {
+		return CreateEmptyClusterValue(), err
+	}
+	data, err := handleInterfaceResponse(res)
+	if err != nil {
+		return CreateEmptyClusterValue(), err
+	}
+	return CreateClusterValue(data), nil
+}